requirements.txt
# requirements.txt
fastapi
uvicorn
pytest
requests
pydantic
httpx
email-validator
python-dotenv==1.0.1